+++ app/actions/memberActions.js
@@ -17,6 +17,7 @@ import {
 } from '../constants/messageConstants'
 
 import dmapi from '../services/apiClient'
+import { actions } from 'react-redux-form'
 
 export function handleDataLoading(bool) {
   return {
@@ -89,6 +90,7 @@ export function loadSignup(token) {
       if (response.error) {
         dispatch(setErrorMessage(REGISTRATION_NOT_FOUND))
       } else {
+        dispatch(actions.load('createUserForm', response.data))
         dispatch(handleLoadSignupSuccess(response))
       }
       dispatch(handleDataLoading(false))
